Refactoring the UI tests and add a new one
